refactor(finyk): decompose Overview.jsx into sub-components
Overview.jsx був 1129 LOC монолітом — один файл з 12+ візуальними секціями,
десятками useMemo-ланцюжків та inline-JSX. Rerender у корені Overview
перераховував геть усе, навіть коли мінявся лише, наприклад, plannedFlows.

Винесено у src/modules/finyk/pages/overview/ 12 sub-компонентів + один
helper:
- FirstInsightBanner — одноразова підказка для активації
- HeroCard — градієнтна картка з нетворсом
- IncomeExpensePills — пара карток дохід/витрати
- NavButtons — швидкі посилання 'Операції' / 'Бюджети'
- MonthPulseCard — місячна картка з прогнозом і фінпульсом
- NetworthSection — графік нетворсу або placeholder
- BudgetAlertsList — алерти про перевищення бюджетів
- PlanFactCard — план vs факт
- QuickAddCard — персоналізовані швидкі дії
- PlannedFlowsCard — найближчі платежі
- CategoryChartSection — chart категорій витрат або empty state
- FlowRow (memo) — рядок планового потоку
- pulseStyle.js — чиста функція computePulseStyle()

Overview.jsx тепер 522 LOC і відповідає тільки за:
- отримання даних з mono/storage
- useMemo-селектори й flow-розрахунки
- композицію sub-компонентів (JSX-лейаут)

Поведінка та стилі не змінились — це чисто структурний refactor.

Тести: 543/543 passing. Lint/typecheck clean.
