+++ src/modules/finyk/pages/overview/CategoryChartSection.jsx
@@ -1,0 +1,55 @@
+import { Suspense } from "react";
+import { CategoryChart } from "../../components/charts/lazy";
+import { ChartFallback } from "../../components/charts/ChartFallback";
+import { EmptyState } from "@shared/components/ui/EmptyState";
+
+/**
+ * «Витрати за категоріями». Відмальовує chart якщо є дані, або empty-state
+ * з CTA на «Операції». Клік по bar-у викликає onCategoryClick + перехід.
+ */
+export function CategoryChartSection({
+  catSpends,
+  onNavigate,
+  onCategoryClick,
+}) {
+  if (catSpends.length > 0) {
+    return (
+      <div className="bg-panel border border-line/60 rounded-2xl p-5 shadow-card">
+        <div className="text-xs font-medium text-subtle mb-4">
+          Витрати за категоріями
+        </div>
+        <Suspense fallback={<ChartFallback className="h-40" />}>
+          <CategoryChart
+            data={catSpends.slice(0, 6)}
+            onBarClick={
+              onCategoryClick
+                ? (catId) => {
+                    onCategoryClick(catId);
+                    onNavigate?.("transactions");
+                  }
+                : undefined
+            }
+          />
+        </Suspense>
+      </div>
+    );
+  }
+
+  return (
+    <div className="bg-panel border border-dashed border-line/60 rounded-2xl shadow-card">
+      <EmptyState
+        title="Поки немає витрат"
+        description="Цього місяця витрат за категоріями ще немає."
+        action={
+          <button
+            type="button"
+            onClick={() => onNavigate("transactions")}
+            className="text-sm font-medium text-primary hover:underline"
+          >
+            Переглянути операції
+          </button>
+        }
+      />
+    </div>
+  );
+}
